+++ spec/rubocop/cop/rspec/named_subject_spec.rb
@@ -62,8 +62,8 @@ RSpec.describe RuboCop::Cop::RSpec::NamedSubject, :config do
     end
   end
 
-  context 'when AllowInSharedExamples is false' do
-    let(:cop_config) { { 'AllowInSharedExamples' => false } }
+  context 'when IgnoreSharedExamples is false' do
+    let(:cop_config) { { 'IgnoreSharedExamples' => false } }
 
     it_behaves_like 'checking subject outside of shared examples'
 
@@ -93,8 +93,8 @@ RSpec.describe RuboCop::Cop::RSpec::NamedSubject, :config do
     end
   end
 
-  context 'when AllowInSharedExamples is true' do
-    let(:cop_config) { { 'AllowInSharedExamples' => true } }
+  context 'when IgnoreSharedExamples is true' do
+    let(:cop_config) { { 'IgnoreSharedExamples' => true } }
 
     it_behaves_like 'checking subject outside of shared examples'