+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## Master (Unreleased)
 
+* Add `IgnoreSharedExamples` option for `RSpec/NamedSubject`. ([@RST-J][])
+
 ## 1.30.1 (2018-11-01)
 
 * `FactoryBot/CreateList` now ignores `times` blocks with an argument. ([@Darhazer][])
-* Add `AllowInSharedExamples` option for `RSpec/NamedSubject`. ([@RST-J][])
 
 ## 1.30.0 (2018-10-08)